+++ CHANGELOG.md
@@ -14,7 +14,7 @@ All notable changes to this project will be documented in this file.
 -->
 
 ### Fixed
-- Centers map and changes zoom level after using locate-me or pick location on map
+- Centers map and changes zoom level after using locate-me
 
 ## [4.6.0] - 2020-10-23