+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - Centers map and changes zoom level after using locate-me
+- Changes picker position to match cursor 
 
 ## [4.6.0] - 2020-10-23